+++ Code/Components/CP/ingest/current/tests/configuration/FeedConfigTest.h
@@ -62,7 +62,7 @@ class FeedConfigTest : public CppUnit::TestFixture {
             casa::Double dblTolerance = 1e-15;
 
             // Create an instance to test
-            const casa::Double nFeeds = 3;
+            const casa::Int nFeeds = 3;
             casa::Matrix<casa::Quantity> offsets(nFeeds, 2);
             casa::Vector<casa::String> pols(nFeeds);
             for (casa::Int i = 0; i < nFeeds; ++i) {
@@ -82,7 +82,7 @@ class FeedConfigTest : public CppUnit::TestFixture {
         }
 
         void testExceptions() {
-            const casa::Double nFeeds = 3;
+            const casa::Int nFeeds = 3;
             {
                 casa::Matrix<casa::Quantity> offsets(nFeeds, 2);
                 casa::Vector<casa::String> pols(nFeeds+1);
